% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{chisq_function}
\alias{chisq_function}
\title{Chi-square function}
\usage{
chisq_function(table)
}
\arguments{
\item{table}{\itemize{
\item data for chi.sq to process
}}
}
\value{
chisq
}
\description{
function to perform chi square test with following options:
correct=FALSE - don't apply continuity correction when computing the test statistic for 2 by 2 tables:
simulate.p.value = true - compute p-values by Monte Carlo simulation.
}
