% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cs_es_table.R
\name{create_cs_es_table}
\alias{create_cs_es_table}
\title{Create cs_es_table}
\usage{
create_cs_es_table(df1)
}
\arguments{
\item{df1}{\itemize{
\item main df1
}}
}
\value{
\itemize{
\item cs_es_table
}
}
\description{
Task 1 - distribution of the expected net income (se)
in relation to the current net income (cs)
}
