% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_es_ab_table.R
\name{create_es_ab_table}
\alias{create_es_ab_table}
\title{Create es_ab_table}
\usage{
create_es_ab_table(df1)
}
\arguments{
\item{df1}{\itemize{
\item dataset for processing
}}
}
\value{
es_ab_table
}
\description{
Task 2 - correlations desire to earn more (se)
between possibility to commute longer (ab)
}
