% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_preprocess.R
\name{general_data_preprocess}
\alias{general_data_preprocess}
\title{general data transformation}
\usage{
general_data_preprocess(df1)
}
\arguments{
\item{df1}{\itemize{
\item input dataframe
}}
}
\value{
df1 - transformed dataframe
}
\description{
general data transformation
}
