% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{ggplot_prep_function}
\alias{ggplot_prep_function}
\title{ggplot table function}
\usage{
ggplot_prep_function(table)
}
\arguments{
\item{table}{\itemize{
\item data for ggplot to process
}}
}
\value{
dt1
}
\description{
function to prepare contingency table data for ggplot2 use
}
