% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{glm_function}
\alias{glm_function}
\title{logistic regression function - for calculating statistics - Task 3 level 2}
\usage{
glm_function(y, x, dataset)
}
\arguments{
\item{y}{\itemize{
\item parameter for glm
}}

\item{x}{\itemize{
\item parameter for glm
}}

\item{dataset}{\itemize{
\item dataset for the function
}}
}
\value{
predictorvariable
}
\description{
logistic regression function for passing x and y parameters plus the dataset
}
