% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{plot_ROC_function}
\alias{plot_ROC_function}
\title{Basic R plot function}
\usage{
plot_ROC_function(plot_dataset, plot_title, plot_subtitle)
}
\arguments{
\item{plot_dataset}{\itemize{
\item dataset for plot
}}

\item{plot_title}{\itemize{
\item title for plot
}}

\item{plot_subtitle}{\itemize{
\item sub title for plot
}}
}
\value{

}
\description{
function for basic R plot with colours and threshold labels
}
