% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{sesp_func}
\alias{sesp_func}
\title{specificity sensitivity function}
\usage{
sesp_func(sesp_variable, predict_variable, threshold)
}
\arguments{
\item{sesp_variable}{\itemize{
\item sensitivity / specificity variable
}}

\item{predict_variable}{\itemize{
\item predictor variable
}}

\item{threshold}{\itemize{
\item threshold level
}}
}
\value{
sensitivity, specificity
}
\description{
specificity sensitivity function
}
