% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task3_preprocessing.R
\name{task3_preprocessing_level1}
\alias{task3_preprocessing_level1}
\title{Task 3 preprocessing level 1}
\usage{
task3_preprocessing_level1(df1)
}
\arguments{
\item{df1}{\itemize{
\item input data frame
}}
}
\value{
df_level1
}
\description{
Preprocessing Data Level 1
}
