% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task3_preprocessing.R
\name{task3_preprocessing_level2}
\alias{task3_preprocessing_level2}
\title{Preprocessing Data Level 2}
\usage{
task3_preprocessing_level2(df1)
}
\arguments{
\item{df1}{\itemize{
\item input data frame
}}
}
\value{
df_level2
}
\description{
put in code to change all NAs to zeros in df1 for logistic regression
}
\details{
convert all NAs in columns curr_sal, exp_sal, hf_com, ab_com to 0
to allow for logistic regression based on factors
}
